+++ hc/api/tests/test_notify_discord.py
@@ -13,7 +13,7 @@ from hc.test import BaseTestCase
 
 
 class NotifyDiscordTestCase(BaseTestCase):
-    def _setup_data(self, value: str, status: str = "down") -> None:
+    def _setup_data(self, url: str, status: str = "down") -> None:
         self.check = Check(project=self.project)
         self.check.name = "foo"
         # Transport classes should use flip.new_status,
@@ -29,7 +29,7 @@ class NotifyDiscordTestCase(BaseTestCase):
 
         self.channel = Channel(project=self.project)
         self.channel.kind = "discord"
-        self.channel.value = value
+        self.channel.value = json.dumps({"webhook": {"url": url}})
         self.channel.save()
         self.channel.checks.add(self.check)
 
@@ -40,8 +40,7 @@ class NotifyDiscordTestCase(BaseTestCase):
 
     @patch("hc.api.transports.curl.request", autospec=True)
     def test_it_works(self, mock_post: Mock) -> None:
-        v = json.dumps({"webhook": {"url": "https://example.org"}})
-        self._setup_data(v)
+        self._setup_data("https://example.org")
         mock_post.return_value.status_code = 200
 
         self.channel.notify(self.flip)
@@ -59,8 +58,7 @@ class NotifyDiscordTestCase(BaseTestCase):
 
     @patch("hc.api.transports.curl.request", autospec=True)
     def test_it_rewrites_discordapp_com(self, mock_post: Mock) -> None:
-        v = json.dumps({"webhook": {"url": "https://discordapp.com/foo"}})
-        self._setup_data(v)
+        self._setup_data("https://discordapp.com/foo")
         mock_post.return_value.status_code = 200
 
         self.channel.notify(self.flip)
@@ -73,8 +71,7 @@ class NotifyDiscordTestCase(BaseTestCase):
 
     @patch("hc.api.transports.curl.request", autospec=True)
     def test_it_handles_last_ping_failure(self, mock_post: Mock) -> None:
-        v = json.dumps({"webhook": {"url": "123"}})
-        self._setup_data(v)
+        self._setup_data("123")
         mock_post.return_value.status_code = 200
 
         self.ping.kind = "fail"
@@ -86,3 +83,14 @@ class NotifyDiscordTestCase(BaseTestCase):
         attachment = mock_post.call_args.kwargs["json"]["attachments"][0]
         fields = {f["title"]: f["value"] for f in attachment["fields"]}
         self.assertEqual(fields["Last Ping"], "Failure, 10 minutes ago")
+
+    @patch("hc.api.transports.curl.request", autospec=True)
+    def test_it_disables_channel_on_404(self, mock_post: Mock) -> None:
+        self._setup_data("https://example.org")
+        mock_post.return_value.status_code = 404
+
+        self.channel.notify(self.flip)
+        # Make sure the HTTP request was made only once (no retries):
+        self.assertEqual(mock_post.call_count, 1)
+        self.channel.refresh_from_db()
+        self.assertTrue(self.channel.disabled)