+++ hc/api/tests/test_notify_slack.py
@@ -158,17 +158,6 @@ class NotifySlackTestCase(BaseTestCase):
         self.assertEqual(n.error, "Slack notifications are not enabled.")
 
     @patch("hc.api.transports.curl.request", autospec=True)
-    def test_it_does_not_retry_404(self, mock_post: Mock) -> None:
-        self._setup_data("123")
-        mock_post.return_value.status_code = 404
-
-        self.channel.notify(self.flip)
-
-        n = Notification.objects.get()
-        self.assertEqual(n.error, "Received status code 404")
-        self.assertEqual(mock_post.call_count, 1)
-
-    @patch("hc.api.transports.curl.request", autospec=True)
     def test_it_disables_channel_on_404(self, mock_post: Mock) -> None:
         self._setup_data("123")
         mock_post.return_value.status_code = 404