+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Implement dynamic favicon in the projects overview page (#971)
 - Add support for system theme (#978, @moraj-turing)
 - Improve Opsgenie notifications (include description, schedule, link etc.)
+- Update the Discord integration to disable channel on HTTP 404 responses
 
 ### Bug Fixes
 - Fix hc.front.views.docs_search to handle words "AND", "OR", "NOT" as queries